# feat(perf): update contacts filter query